# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget control API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/pss_budget_control/v1/categories/store": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create category item",
                "description": "Create category item.",
                "operationId": "createCategoryItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "type_id",
                                    "name"
                                ],
                                "properties": {
                                    "type_id": {
                                        "description": "Category type ID (1 - incomes, 2 - outcomes)",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Category name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pss_budget_control/v1/categories/{category}/update": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category item",
                "description": "Update category item.",
                "operationId": "updateCategoryItem",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "type_id",
                                    "name"
                                ],
                                "properties": {
                                    "type_id": {
                                        "description": "Category type ID (1 - incomes, 2 - outcomes)",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Category name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Wrong category ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pss_budget_control/v1/categories/{category}/delete": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category item",
                "description": "Delete category item.",
                "operationId": "deleteCategoryItem",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Wrong category ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pss_budget_control/v1/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category item",
                "description": "Get category item.",
                "operationId": "getCategoryItem",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CategoryItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong category ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pss_budget_control/v1/categories/list": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories list",
                "description": "Get categories list.",
                "operationId": "getCategoriesList",
                "parameters": [
                    {
                        "name": "search_string",
                        "in": "query",
                        "description": "Search string for category name",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Example category"
                    },
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "Category type ID",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "items_per_page",
                        "in": "query",
                        "description": "Amount of categories for limit",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    },
                    {
                        "name": "items_skip",
                        "in": "query",
                        "description": "Amount of categories for skip",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    },
                    {
                        "name": "order_by_direction",
                        "in": "query",
                        "description": "Categories ordering direction",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "DESC or ASC"
                    },
                    {
                        "name": "order_by_field",
                        "in": "query",
                        "description": "Categories ordering fied",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "id, name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryItem"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryItem": {
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Category ID",
                        "type": "integer",
                        "default": "1"
                    },
                    "type_id": {
                        "title": "Type ID",
                        "description": "Category type ID",
                        "type": "integer",
                        "default": "1-2"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Category name",
                        "type": "string",
                        "default": "Some name"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoryItem"
                }
            },
            "ErrorResource": {
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Error message",
                        "type": "string",
                        "default": "Not found"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Error"
                }
            },
...
            "ValidationErrorResource": {
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Error message",
                        "type": "string",
                        "default": "Unprocessable Entity"
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors array",
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ValidationError"
                }
            }
        }
    }
}